util: Remove salt from secure_id, instead add .
The rationale for this change is to obtain a static permalink for Games
and Sheets, even if the link is recreated by calling encode_secure_id()
again. However, we don't want links to be guessable from other known
links, so we add the `realm` to the hash to make secured Game and Sheet
ids disjoint.