Add detailed user names (first_name, last_name, username instead of name)
This changes the database schema to store the full Telegram user name
information. Users need to issue the private `/start` command again to
get their names in an existing database updated. To access the full name
in the code easily, the `model.User.format_name()` method is added. It
allows to retreive the full name or a short version (only first name),
which can optionally be made unambiguous by providing a name of
conflicting users.

Fixes #19